:bug: 主键修改为 Long 类型遗漏部分提交

+++ ballcat-common/ballcat-common-util/src/test/java/com/hccake/ballcat/common/util/TreeUtilsTest.java
@@ -20,15 +20,15 @@ class TreeUtilsTest {
 	 * @param levelNums 每个节点的子节点数量
 	 * @return List
 	 */
-	private List<TreeNode<Integer>> buildTreeNodes(List<Integer> levelNums) {
-		List<TreeNode<Integer>> list = new ArrayList<>();
-		List<Integer> parentIds = new ArrayList<>(levelNums.size());
-		parentIds.add(0);
+	private List<TreeNode<Long>> buildTreeNodes(List<Integer> levelNums) {
+		List<TreeNode<Long>> list = new ArrayList<>();
+		List<Long> parentIds = new ArrayList<>(levelNums.size());
+		parentIds.add(0L);
 
-		int id = 0;
+		long id = 0L;
 		for (int i = 0; i < levelNums.size(); i++) {
 			int levelNum = levelNums.get(i);
-			int parentId = parentIds.get(i);
+			Long parentId = parentIds.get(i);
 
 			for (int j = 0; j < levelNum; j++) {
 				id++;
@@ -46,29 +46,29 @@ class TreeUtilsTest {
 	@Test
 	void treeTest() {
 		// 构建一个树节点列表
-		List<TreeNode<Integer>> list = buildTreeNodes(Arrays.asList(2, 1, 1, 2, 3, 4));
+		List<TreeNode<Long>> list = buildTreeNodes(Arrays.asList(2, 1, 1, 2, 3, 4));
 		// 树节点列表转树结构
-		List<TreeNode<Integer>> nodes = TreeUtils.buildTree(list, 0);
+		List<TreeNode<Long>> nodes = TreeUtils.buildTree(list, 0L);
 		// 树结构转回树节点列表
-		List<TreeNode<Integer>> abstractIdTreeNodes = TreeUtils.treeToList(nodes);
+		List<TreeNode<Long>> abstractIdTreeNodes = TreeUtils.treeToList(nodes);
 		// 排序处理
-		abstractIdTreeNodes.sort(Comparator.comparingInt(TreeNode::getKey));
+		abstractIdTreeNodes.sort(Comparator.comparingLong(TreeNode::getKey));
 
 		// 比较初始的树节点列表和通过两次转换完毕的结果是否一致
 		Assertions.assertEquals(list, abstractIdTreeNodes);
 	}
 
-	static class TestTreeNode implements TreeNode<Integer> {
+	static class TestTreeNode implements TreeNode<Long> {
 
 		/**
 		 * 节点ID
 		 */
-		private Integer id;
+		private Long id;
 
 		/**
 		 * 父节点ID
 		 */
-		private Integer parentId;
+		private Long parentId;
 
 		/**
 		 * 子节点集合
@@ -76,40 +76,40 @@ class TreeUtilsTest {
 		private List<TestTreeNode> children;
 
 		@Override
-		public Integer getKey() {
+		public Long getKey() {
 			return id;
 		}
 
 		@Override
-		public Integer getParentKey() {
+		public Long getParentKey() {
 			return parentId;
 		}
 
 		@Override
 		@SuppressWarnings("unchecked")
-		public <T extends TreeNode<Integer>> void setChildren(List<T> children) {
+		public <T extends TreeNode<Long>> void setChildren(List<T> children) {
 			this.children = (List<TestTreeNode>) children;
 		}
 
 		@Override
 		@SuppressWarnings("unchecked")
-		public <T extends TreeNode<Integer>> List<T> getChildren() {
+		public <T extends TreeNode<Long>> List<T> getChildren() {
 			return (List<T>) children;
 		}
 
-		public Integer getId() {
+		public Long getId() {
 			return id;
 		}
 
-		public void setId(Integer id) {
+		public void setId(Long id) {
 			this.id = id;
 		}
 
-		public Integer getParentId() {
+		public Long getParentId() {
 			return parentId;
 		}
 
-		public void setParentId(Integer parentId) {
+		public void setParentId(Long parentId) {
 			this.parentId = parentId;
 		}
